moves the tests along the component
