+++ src/signals/incident-management/components/CalendarInput/index.ts
@@ -1,0 +1,3 @@
+import CalendarInput from './CalendarInput';
+
+export default CalendarInput;
